requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.4
omegaconf==2.3.0
openai-whisper==20240930
openunmix==1.3.0
packaging==24.2
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
proglog==0.1.11
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.2
submitit==1.5.2
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
treetable==0.2.5
typing_extensions==4.13.2
urllib3==2.4.0
